2D graphs for Dirichlet and Normal doc: PDF & CDF
Closes #1052.

PDFs of 2 different Dirichlet distributions
are plotted in addition to the existing 1D PDF and CDF graphs.

The plots of the PDF and CDF of a 1D normal distribution
are replaced with the PDF and CDF of a 2D normal distribution.

Distributions with a 2D PDF isolines plot also receive a 2D CDF isolines
plot.